requirements.txt
flask==0.12.2
flask_sqlalchemy==2.3.2
flask_login==0.4.1
psycopg2-binary==2.7.4
flask_wtf==0.14.2
bcrypt==3.1.4
flask-restful==0.3.6
flask_migrate==2.1.1
flask_jwt_extended==3.8.1
flask_httpauth==3.2.3
flask-script==2.0.6